+++ src/components/settings/index.js
@@ -1,8 +1,24 @@
 import React, { Component, PropTypes } from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
+import { reduxForm } from 'redux-form'
 import * as Actions from '../../actions/auth'
 
+const validate = values => {
+  const errors = {};
+  if (!values.nickname) {
+    errors.nickname = 'Required';
+  } else if (!/^[(\u4e00-\u9fa5)0-9a-zA-Z\_\s@]+$/.test(values.nickname)) {
+    errors.nickname = '无效电子邮件地址';
+  }
+  return errors;
+};
+
+@reduxForm({
+  form: 'settings',
+  fields: ['nickname'],
+  validate
+})
 class Settings extends Component {
   constructor(props){
     super(props)
@@ -11,24 +27,25 @@ class Settings extends Component {
   
   handelSubmit(e){
     e.preventDefault();
-    const {nickname} = this.refs
+    const {values} = this.props
     //更新user nickname
     const {actions} = this.props
-    actions.updateUser({nickname:nickname.value})
+    actions.updateUser(values)
   }
-  componentWillReceiveProps(nextProps){
-    const { auth } = nextProps
-    if(auth.errMsg){
-      msg.error(auth.errMsg)
-    }else if(auth.user.nickname !== this.props.auth.user.nickname){
-      msg.success('修改成功.')
+  validatorCalss(field){
+    let initClass = 'form-control'
+    if(field.invalid){
+      initClass += ' ng-invalid'
     }
+    if(field.dirty){
+      initClass += ' ng-dirty'
+    }
+    return initClass
   }
   render() {
-    const { actions,auth } = this.props
+    const { actions,auth,fields: { nickname },dirty,invalid,valid } = this.props
     return (
       <div className="settings-box">
-
         <div className="settings-container">
           <h2 className="title">设置</h2>
           <hr />
@@ -38,13 +55,12 @@ class Settings extends Component {
                   <div className="form-group">
                     <label className="control-label">昵称</label>
                     <input placeholder="2-15字符，中英文、数字和下划线" 
-                          required 
+                          {...nickname}
                           type="text" 
-                          ref="nickname"
-                          defaultValue={auth.user.nickname}
-                          className="form-control" minLength="2" maxLength="15" />
+                          className={ this.validatorCalss(nickname) }
+                          minLength="2" maxLength="15" />
                   </div>
-                  <button type="submit" className="btn btn-block btn-lg btn-primary">保 存</button>
+                  <button type="submit" disabled={ dirty && invalid } className="btn btn-block btn-lg btn-primary">保 存</button>
                 </form>
             </div>
           </div>
@@ -61,7 +77,8 @@ Settings.propTypes = {
 
 function mapStateToProps(state) {
   return {
-    auth: state.auth.toJS()
+    auth: state.auth.toJS(),
+    initialValues: state.auth.toJS().user
   }
 }