+++ src/containers/Home.js
@@ -4,14 +4,13 @@ import { connect } from 'react-redux'
 import Header from '../components/header'
 import * as Actions from '../actions'
 import * as AuthActions from '../actions/auth'
-import Toaster from '../components/toaster'
+import Toaster from '../components/Toaster'
 import ScrollTop from '../components/scrollTop'
 
 class Home extends Component {
   constructor(props){
     super(props)
   }
-
   componentDidMount() {
     const { actions,authActions,auth } = this.props
     actions.getIndexImage()
@@ -25,14 +24,13 @@ class Home extends Component {
       document.body.className = nextProps.globalVal.styleMode
     }
   }
-
   render() {
-    const { globalVal,actions,children,auth,authActions,location } = this.props
+    const { globalVal,actions,children,auth,authActions,location,showmsg } = this.props
     return (
       <div>
         <Header styleMode={globalVal.styleMode} auth={auth} logout={authActions.logout} location={location} changeStyleMode={actions.changeStyleMode} />
         {children}
-        <Toaster />
+        <Toaster msg={showmsg} hideMsg={authActions.hideMsg} />
         <ScrollTop />
       </div>
     )
@@ -42,6 +40,7 @@ class Home extends Component {
 Home.propTypes = {
   globalVal: PropTypes.object.isRequired,
   auth: PropTypes.object.isRequired,
+  showmsg: PropTypes.object.isRequired,
   children: PropTypes.node,
   actions: PropTypes.object.isRequired
 }
@@ -49,7 +48,8 @@ Home.propTypes = {
 function mapStateToProps(state) {
   return {
     globalVal: state.globalVal.toJS(),
-    auth: state.auth.toJS()
+    auth: state.auth.toJS(),
+    showmsg: state.showmsg.toJS(),
   }
 }