+++ src/store/configureStore.js
@@ -1,13 +1,15 @@
 import { createStore,compose,applyMiddleware } from 'redux'
-import rootReducer from '../reducers'
 import thunkMiddleware from 'redux-thunk'
+import { routerMiddleware, push } from 'react-router-redux'
+import { Router, Route, browserHistory } from 'react-router'
 import {persistState} from 'redux-devtools'
 import DevTools from '../containers/DevTools'
+import rootReducer from '../reducers'
 //使用chrome 扩展来做调试工具.
 // window.devToolsExtension ? window.devToolsExtension() : f => f
 
 let finalCreateStore
-const middleware = applyMiddleware(thunkMiddleware)
+const middleware = applyMiddleware(routerMiddleware(browserHistory),thunkMiddleware)
 if(process.env.NODE_ENV === 'production'){
   finalCreateStore = compose(middleware)
 }else{