+++ History.md
@@ -1,3 +1,14 @@
+1.4.0 / 2016-03-16
+==================
+
+* 更新redux-simple-router 到 react-router-redux
+* 更新redux 到 3.3.1
+* 更新react 到 0.14.7
+* 更新react-router 到 2.0.1
+* 弃用react-alert 使用 react-s-alert
+* 重新设计消息提示组件.
+
+
 1.3.1 / 2016-02-28
 ==================